backend/ingest/requirements.txt
boto3
faiss-cpu
requests
python-dotenv
